Change has_require_members to singular has_require_member.
